drv-rng-api: Add new client library to get entryopy from RNG.
This API is intended to simplify getting entropy from the rng_driver. It
exposes two functions:

rng_fill - a function that takes the TaskId of the RNG server and a
slice that the function will fill with entropy for the caller

rng_getrandom - a function compatible with the 'getrandom' crate custom
entropy feature, it takes a slice that the function will fill in with
entropy

Compatibilty with the 'getrandom' crate is protected by the local
'custom-getrandom' feature that is disabled by default. If you're
implementing an rng_driver you can use the api crate in its default
config to get common error codes through the RngError enum.

The ugly part of this crate is the 'rng_getrandom' function where I've
not found a clean way to provide the RNG driver TaskId to the custom
'getrandom' function. In C I'd provide an init function that callers can
use to populates a static variable for the TaskId in the library. I've
no clue yet how to accomplish similar things with Rust so this currently
uses the task_slot! macro. The implications of this choice aren't yet
clear to me but it works.