drv-lpc55-rng: Change interface to fill an arbitrary buffer with entropy
The RNG hardware on the lpc55 provides a 32bit register holding the
entropy generated by the platform hardware. The previous interface would
read a u32 from this register and return it the caller. The rust rand
crate recommends using a 32 byte seed for CPRNGs[1] and making 8
syscalls to get this much entropy seems inefficient. As an alternative
this commit changes the lpc55-rng syscall to fill an arbitrary buffer
provided by the caller.

[1] https://docs.rs/rand/0.5.0/rand/trait.SeedableRng.html