lpc55-rng: Update interface to fill an arbitrary buffer with entropy.
The RNG hardware on the lpc55 provides a 32bit register holding the
entropy generated by the platform hardware. The previous interface would
read a u32 from this register and return it the caller. The rust rand
crate recommends using a 32 byte seed for CPRNGs[1] and making 8
syscalls to get this much entropy seems inefficient.

This commit changes the lpc55-rng syscall to fill an arbitrary buffer
provided by the caller. This new interface is defined using the IDL.

To simplify access to the RNG we add the rng-api library crate that:
1) wraps the new syscall in the RngCore trait from the 'rand_core' crate
2) includes a custom 'getrandom' implementation gated by the
'custom-getrandom' feature

This API crate allows consumers to use the familiar 'getrand' and
'rand::os::OsRng' interfaces to get random bytes from the rng_server
task.

[1] https://docs.rs/rand/latest/rand/trait.SeedableRng.html