lpc55-rng: Condition RNG output with reseeding ChaCha20 rng.
Chacha20 rng will be reseeded after a threashold number of bytes. This
is currently set to 1 MiB. This is a similar approach to the one taken
in rand::rngs::adapter::ReseedingRng but without the std dependency that
comes from dealing with process forking.

The ChaCha20Core type used is from the rand_chacha crate. This uses ~300
bytes of stack more than ChaCha20RngCore from the chacha20 crate but the
end result uses ~6K less flash.