wip: disable ringbuf in hiffy stm32h7 module, enable in common
This was necessary for some debugging. Hiffy calls into the RNG to get
entropy that it hands up to humility. This is done through the IDL.
Unfortunately rand_core wants us to start our error codes at
Error::CUSTOM_START that turns out to be a large number. When the IDL
serializes this large number it either wraps or is truncated. In any
case it becomes 0 and is interpreted as success.

Debugging this requred enabling ringbuf in hiffy common, which required
disabling the ringbuf in the stm32h7 module. I attempted to name the
ringbufs so the they could coexist but this didn't work.

None of this should be committed.