lib-lpc55-usart: Add library to wrap access uart device on lpc55.
This library implements the embedded-hal uart traits for nonblocking
read and write. This uses the 'nb' crate for appropriate error
responses.

lib-lpc55-usart includes a 'hubris' feature that's enabled by default.
This feature controls whether or not the library uses supporting tasks
to configure GPIOs, clocks etc. If disabled the library assumes that it
is able to access these hardware regions directly and will configure
them using the lpc55-pac crate.