lib-dice: Add mfg interface, move common types to mfg lib.
New mfg lib provides types for blobs like the CSR and Certs where one end
of the mfg communication channel can't know the appropriate size in
advance. Specifically the mfg side won't know how big the CSR is, and
the RoT won't know the size of the DeviceId or intermediate certs. The
platform serial number is part of the API for the mfg commands so that's
been moved into this library as well.