sprot: plumbing for attest & attest_len
This is the first op supported by sprot that takes both a read and a
write lease. Following the existing structure of `handle_request` we add
a new member to the TrailingData enum to hold a slice with the data in
the read lease. As a result we must be more explicit about the lifetimes
in the `handle_request` method because it's return value no longer has
the same lifetime as the `self` param (3rd lifetime elision rule).