wip: Implement 'attest'
This will probably "work" but it's not something we can merge. The issue
this commit hasn't resolved is that the existing code doesn't allow us
to access the data that is our nonce in the code that handles the
`TrailingData` returned by `handle_request`. By the time we get there
the `Request` struct has gone out of scope and that data is no longer
available.