lpc55-rng: Include DICE derived seed in initial PRNG seed.
Previously we constructed the initial seed as 32 bytes from the hardware
RNG:

```
SEED_0 = HRNG(32)
```

This commit includes a seed value constructed by the measured boot
implementation from the DICE CDI. This is passed through to the RNG task
using the `stage0-handoff` mechanism. This 32 byte value is now
extracted and mixed with 32 bytes from the HRNG to construct SEED_0:

```
SEED_0 = sha3_256(DICE_SEED | HRNG(32))
```

Use of this feature is gated by the `dice-seed` feature that is enabled
by default. Images that do not want to use this seeding strategy (when
DICE is disabled) will have to disable it explicitly.