wip: this seems to effect on the resultant idle binary
After we stopped explicitly adding `--no-default-features` to the rustc
command manually in xtask I noticed that the binary for the idle task
was changed when I had expected that it wouldn't be. This *appears* to
have been caused by an empty set of default features in the idle task
Cargo.toml. No clue why this causes the idle task binary to change while
not changing the final image.